integration tests: move status tests to separate file
